【SYKFPT-1093】
Fix clippy warnings across workspace, restore mssql driver, and resolve compilation errors
- Re-implemented MssqlBuilder and SchemaMigration for MSSQL
- Cleaned up clippy warnings (matches!, clamp, ptr_arg, FFI ptr deref safety)
- Added docker-compose-mssql.yml for local E2E MSSQL testing
- Fixed compilation error regarding std::path::Path

+++ cli/cowen/crates/cowen-server/src/daemon/manager.rs
@@ -114,7 +114,7 @@ impl WorkerManager {
                             }
                         } else {
                             // Exponential backoff with 10% jitter logic
-                            let base_delay = 2u64.pow(next_retry as u32).min(60) as f64;
+                            let base_delay = 2u64.pow(next_retry).min(60) as f64;
                             use rand::Rng;
                             let jitter = rand::thread_rng().gen_range(0.0..=0.2);
                             let delay_secs = base_delay * (0.9 + jitter);
